parseUserAgent function
I needed this for a project and was not satisfied with what I found on
the web (not an exhaustive search, of course).  This is a much simpler
function that seems to work on most modern browser user-agent strings.
Of course, user agent parsing is NEVER recommended for feature
detection.  This function was used becausae a crappy 3rd party tool that
was opened in an iframe was sensitive (did noit work) in all versions of
all browser, so I needed something to detemine if the code was going to
work before I launched it.